Propose initial implementation of comment creation
Generally comments has to be linked to a link.

Since it will be tricky to pass around URLs on API URL, so it is
recommended that the client SDK normalizes and hashes out the URL before
making request to the endpoint. Basically "link_id" is the hash of the
normalized version of the URL.